Intercept RenameRepository calls in Praefect
Praefect currently handles RenameRepository like every other mutator.
This is problematic as Praefect ends up renaming repositories first on
the disks of the Gitalys and then in the database. If only the first
operation succeeds, Praefect has effectively lost track of the repositories.
With Praefect now generating unique relative paths for each repository,
there's no need to actually move the repositories on the disks anymore.
It is sufficient to rename the repository only in the database and leave the
replicas in their existing locations on disk. This commit intercepts
RenameRepository in Praefect and renames repositories only in the database.

The atomic rename handling will be rolled out progressively with Praefect
generated replica paths as the change does not make sense on its own.
To do so, the RenameRepositoryFeatureFlagger checks the replica path of
the repository and only applies the new logic if the repository has a
Praefect generated replica path.

Changelog: fixed