Use Praefect's RemoveRepository from remove-repository subcommand
The RemoveRepository subcommand is working around the lack of atomicity
in Praefect's repository removals. It manually cleaned up database state
and issued deletes Gitalys to ensure deletes go through even if there
is no metadata for the repository. This ensured there would be no conflicts
when recreating a repository after running the command.

These workarounds are no longer necessary as Praefect is now handling
deletions. As soon as the repository's metadata is deleted, it stops
existing from Praefects point of view. New repository created in the same
virtual storage and relative path lands in a different directory on the
Gitalys. As such, we can simplify the command and simply call Praefect's
RemoveRepository to delete the repository's metadata and to delete the
known replicas.

The command was changed to error if the repository does not exist as it
can't do anything if so. Likewise, the output was changed to not mention
the nodes the repository exists on as it's mostly irrelevant. The deletion
may succeed as far as the user cares but some replicas may still be left
on the disk if Praefect was unable to remove them.

Changelog: changed