git2go: Fix comments in MergeCommand
In Git, "ours" and "theirs" are named from the perspective of the
maintainer accepting the changes. Ours is the main target ref that
we merge into, and theirs is the source ref.
Based on orders of parents in call to CreateCommitFromIds
(ours, theirs) that is how we use those as well, only this comment
was wrong.
This is probably related to
https://gitlab.com/gitlab-org/gitaly/-/merge_requests/2868
and it was fixed in code, but the mistaken comment was left.
This is important to iron out before we add squashing,
as the wrong parent for squashed commit could be catastrophic.