ci: Add job to exercise minimum required Git version
With the conversion of our CI jobs to use bundled Git by default we're
not testing the minimum required Git version anymore. Add a new job
which tests that we're still compatible with this version to assert that
we're not starting to use new Git features which aren't available in
this version.