repository: Structured errors for UserRebaseConfirmable
Currently, UserRebaseConfirmable will return gRPC OK even when there is
an error during rebase or if there is an error during PreReceive when it
calls the GitLab API for an access check.

This change modifies the error handling to return a detailed gRPC
structured error in either of these failure modes.

Changelog: changed