Makefile: Upgrade go-licenses to v1.0.0
Upgrade go-licenses to v1.0.0. Most importantly, this release contains a
fix to the way project sources are copied into the target directory so
that we don't end up with files and directories which cannot be modified
[1]. The workaround we have in our codebase which fixes the permissions
can be removed as a consequence.

Note that this causes us to pick up an additional file that's part of
our own sources as licensed file because it's a test verifying that
licenses can be parsed alright, and thus it contains a license body
itself. It's known though that go-licenses behaves weird already though:
it copies the complete sourcecode of projects into the target directory
in case it detects that the license requires it. So adding one more
weird edge case to our NOTICE file doesn't really matter. Together with
the fact that we can remove the workaround we previously had this seems
like an acceptable tradeoff.

[1]: https://github.com/google/go-licenses/issues/11