ci: Increase test coverage of PgBouncer job
The PgBouncer CI job is currently only executing the `test-postgres`
target, which only exercises a subset of our test suite. It's much more
sensible thoguh to instead execute `test-with-praefect` so that we can
assert that a "real" Praefect proxy works alright in combination with
PgBouncer.

Do this conversion and remove the now-unused `test-postgres` target.
Convert the job to run tests as unprivileged user so that we don't
ignore permissions, or otherwise some of our tests would break.