maintenance: Call housekeeping manager's OptimizeRepository directly
Currently, OptimizeRepository is called through an RPC to the gitaly
instance. This is unnecessarily roundabout, however. We can just call
the method directly.

This commit adds an OptimizerFunc that makes it easy for a caller to
pass in a function to act as an Optimizer. We then pass in a direct call
to OptimizeRepository from main.go.

Changelog: changed