proto: Mark related RPCs as maintenance operations
We've got all the infrastructure in place now to handle maintenance
operations specially in our infrastructure. Mark all RPCs which are
maintenance-related as maintenance operations to switch them to use the
new infrastucture.

Note that we've still got a feature flag which toggles between old and
new routing behaviour for these RPCs.

Changelog: changed