maintenance: Add metrics to keep track of repository optimizations
When we called the OptimizeRepository RPC, there were some metrics we
got for free through the gRPC middleware. Now that we're not calling the
RPC, we will need to log metrics on our own.

This change adds a metric to keep track of successful and failed calls
to OptimizeRepository.

ff2f4ee8 (housekeeping: Track total number of optimization calls
2022-03-18) added a metric to keep track of the total number of
optimization calls. We also already track the latency of an
OptimizeRepository call. This makes the similiar metrics in
internal/gitaly/maintenance redundant.

Remove the metrics that track OptimizeRepository in
internal/gitaly/maintenance.

In order to have visibility into success and failures for housekeeing
tasks, add a label to the gitaly_housekeeping_tasks_total metric.

In optimizeRepository, there is an anonymous struct to keep track of
whether or not a given housekeeping task ran. This added layer of
indirection doesn't help much, however, since we end up mapping it back
to strings anyways when emitting the metric.

Instead, use a plain map keyed on the metric name with a boolean value
for whether or not it succeeded. This way, if the metric exists in the
map, then we know that it ran and whether or not it succeeded. If a
metric does not exist in the map, then that means the task was never
attempted.