housekeeping: Fix Prometheus metrics' latency buckets
The housekeeping manager is hosting a latency metric which tracks how
long specific tasks of the housekeeping manager take. But while latency
metrics should typically use the GRPC latency buckets as configured in
the Gitaly configuration, we accidentally didn't in the context of the
housekeeping manager.

Fix this bug by passing in the Prometheus configuration.