limithandler: Pave the way for a second limiter type
A future commit will add a new middleware that will limit based on the
rate rather than concurrent calls. There is a good amount of logic
currently used by the concurrency limiter that can be reused since a
rate limiter is also operating on incoming requests based on RPC name.

To make easier to add this new limiter type in the future, refactor the
code by adding some abstractions easier to add another type of limiter.