limithandler: Introduce rate limiter middleware
Introduce a simple rate limiter that limits the number of requests a
minute that an RPC can allow.

If the feature flag is enabled, the middleware will drop any request
that bursts the per second limit of the RPC. Otherwise, it will only emit
metrics so we can first have some data on the traffic profile.