git: Fix bundled Git test relying on state of feature flag
One of our tests for bundled Git verifies that the command factory is
setting up the execution environment as expected. This test implicitly
relies on the state of feature flags though because it assumes that all
bundled Git constructors are usable. This worked alright by pure chance:
the feature flag for this bundled Git constructor is always true because
we use `math.Rand()` to set the flag's value, which isn't ever properly
seeded.

Fix this issue by hard-coding the feature flag value to `true` so that
the test doesn't break when the randomly injected value changes.