localrepo: Add Size() for repo
Add a Size() method for repo that calculates disk usage of objects. The
way we currently calculate repository size is to use du(1), which is a
quick and dirty way to get the total storage of a repository. However,
this ends up varying depending on the __how__ the objects are stored by
Git, which is affected by repository maintenance strategies and can vary.

A more consistent way to calculate storage is to get a total of number
of reachable objects with git rev-list --all --objects --disk-usage