housekeeping: Avoid stat calls when searching for locked refs
We use `filepath.Walk()` to search for stale reference locks, which will
stat every single directory entry we're finding. This is needless
busywork though: we would only ever consider files for deletion which
have a `.lock` suffix, so all the other stat calls are wasted.

Convert the function to instead use `filepath.WalkDir()` to reduce the
number of stat calls.

Changelog: performance