housekeeping: Avoid stat calls when searching for temporary objects
We use `filepath.Walk()` to search for temporary objects, which will
stat every single directory entry we're finding. This is needless
busywork though: we only need to stat files which have a `tmp_` prefix.

Convert the function to instead use `filepath.WalkDir()` to reduce the
number of stat calls.

Changelog: performance