Makefile: Run goimports only once
As part of the `format` Makefile target we also run `goimports` to
automatically fix up missing or superfluous imports. Because of the way
`goimports` adds these imports, the end result may not conform to the
code style as dictated by `gofumpt`. To fix any formatting differences,
we call `goimports` twice: once before, and once after `gofumpt`.

This doesn't really make a lot of sense though: executing it once before
the formatter runs should be sufficient. For it to not be sufficient the
formatter would have to remove imports again, and this is something I'd
claim doesn't happen.

Remove the second call to `goimports` to speed up the `format` target.