Makefile: Copy only Gitaly commands in install
In the `install` target all binaries found in _build/bin were installed.
At the same time, the `install-bundled-git` target is a prerequisite of
that `install` target. So this would cause the bundled Git binaries to
be copied twice. That's not needed, so instead only copy the Gitaly
commands in the `install` target.