limithandler: Change metric name for concurrency limiting
Before, concurrency limiting was the only limiting middleware. The name
of the metric had gitaly_rate_limiting in it, which was a bit of a
misnomer since rate was never part of the equation. Now that we are
actually adding a rate limiter to Gitaly, the concurrency metrics will
be mistaken for rate limiting metrics.

Change the name of these by replacing rate_limiting with
concurrency_limiting.

Changelog: changed