internalgitaly: WalkRepos to return last accessed timestamp
In order to make a decision on which repositories to clean up,
Praefect's repocleaner will match repository paths with what it sees in
the database. If it doesn't see a record in the database, that means
Praefect doesn't know about this repository.

However, it could be the case that a repository has just been created
and the record for it doesn't yet exist in the database.

In order for repocleaner to know, it needs the information to be
returned from Gitaly.

Add a field in the RPC response, and return the last modified date on the
refs/ dir so we can use it as a proxy for when the repository was
created.