git: Limit number of threads in git-grep(1)
Similar to git-pack-objects(1), git-grep(1) also spawns as many threads
as there are CPU cores. This has caused multiple incidents already when
there had been multiple invocations of this command in biggish repos
because we start to exhaust CPUs quite fast.

Limit the number of threads for git-grep(1) in the same way as we limit
git-pack-objects(1).

Changelog: performance