operations: Default-enable quarantined voting for UserSquash
We have recently introduced quarantined voting for the UserSquash RPC.
With this change in place, we first stage new objects in a quarantine
directory which only gets merged into the main repository in case the
RPC succeeds. Furthermore, we vote on the generated commits to ensure
that they're the same across all Gitaly nodes in case of a Praefect
setup.

This flag has been rolled out to production on March 24th without any
issues. While the feature seems to be stable enough, we still can't
remove the feature flag though due to backwards compatibility concerns:
all nodes must behave the same already and know how to handle the voting
or otherwise we face problems during zero-downtime upgrades when only
some of the Gitaly nodes have been upgraded yet.

This commit thus default-enables the feature flag before we can remove
it altogether in the next release.

Changelog: changed