housekeeping: Report when we have pruned config keys
We do not currently report when we're pruning config keys. While we
cannot easily derive the number of keys we have pruned, we can at least
report whether we pruned something or not.

Add this reporting so that we can tell whether the config pruning is
still in use or whether all repositories have already been cleaned up.