gittest: Remove "origin" remote in `CloneRepo()`
With the `gittest.CloneRepo()` helper callers can easily create a
repository from a specific source repository. One slight issue with this
helper though is that the resulting repository still has the `origin`
remote after the RPC finishes, which makes it inconsistent with repos
we'd typically see in production.

Let's remove the remote after having cloned the repository. This is a
preparatory commit for an upcoming change where we start to measure how
many config keys we're pruning in `housekeeping.CleanStaleData()`, as
otherwise we'd have to amend all of our testcases in that package to
also report the cleaned-up default repository.