Makefile: Always build Go binaries with a build ID
We have recently introduced support for generating build IDs in Gitaly.
The biggest downside of this support is that we're forced to build all
binaries with separate commands, which is quite a lot slower by default
than when we built all commands with a single execution of the Go
compiler. Because of this we have added a fallback case that allows
developers to skip generation of the build ID so that local development
isn't significiantly slowed down.

Unfortunately, we now have another nail in the coffin for building all
binaries with a single command: we need to be able to specify separate
Go build tags per Go executable. This becomes unworkable though in the
case where we build them all in one go.

Luckily though, Gitaly's Makefile has recently also dropped the
`.NOPARALLEL` special target. With this target gone, we can parallelize
compilation of the different Go binaries via make(1) itself. This brings
us back most of the performance we would loose when we build binaries
with separate invocations of the compiler.

So let's drop the workaround which allowed developers to skip generating
the build ID. Instead, they should simply ask make(1) to parallelize
compilation via the `-j` flag. This change both simplifies our Makefile
and at the same time paves the way for per-binary Go tags.