Makefile: Make libgit2 dependency more granular
Right now the `build` target depends on libgit2, which ultimately means
that we'll only start building Gitaly's binaries when libgit2 has been
compiled already. This is needlessly restrictive though: we have only a
single binary which depends on libgit2, which is `gitaly-git2go`.

Explicitly mark these dependencies such that we can parallelize building
libgit2 and some of our Go binaries.