ssh: Clean up output when pre-receive hook fails
When the pre-receive hook failed, git-receive-pack(1) exits with code 0.
It's arguable whether this is the expected behavior, but anyhow it means
cmd.Wait() did not error out. On the other hand, the gitaly-hooks
command did stop the transaction upon failure. So this final vote fails.

To avoid this error being presented to the end user, ignore it when the
transaction was stopped.

Issue: https://gitlab.com/gitlab-org/gitaly/-/issues/3636
Changelog: fixed