Makefile: Group Git patches by version
The current set of Git patches got quite big, and consequentially it's
hard to see which patches belong to what version. Reorder them into a
per-version subdirectory so that the grouping is clear. Furthermore,
this allows us to find all patches via wildcards instead of having to
manually list them in our Makefile.